fix: team 패키지를 domain 패키징 안으로 이동
